backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.3
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Environment
python-decouple==3.8

# Utilities
python-slugify==8.0.1
Pillow==10.1.0

# Date/Time
pytz==2023.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Scheduled Tasks
celery==5.3.4
redis==5.0.1

# APScheduler for background tasks
apscheduler==3.10.4

# AI / LLM
openai==1.55.0

# Production Server
gunicorn==21.2.0

# Monitoring
sentry-sdk==1.38.0

# Performance
django-redis==5.4.0
